+++ tests/jcli/rest/tip.rs
@@ -1,6 +1,5 @@
 #![cfg(feature = "integration-test")]
 
-use common::configuration::jormungandr_config::JormungandrConfig;
 use common::configuration::node_config_model::NodeConfig;
 use common::jcli_wrapper;
 use common::process_assert;
