--- tests/jcli_genesis/mod.rs
@@ -1,1 +1,0 @@
-pub mod jcli_genesis_test;
