--- tests/jormungandr_config/mod.rs
@@ -1,1 +1,0 @@
-pub mod jormungandr_config_test;
