+++ tests/tests.rs
@@ -1,8 +1,6 @@
 pub mod common;
-pub mod jcli_certificates;
-pub mod jcli_genesis;
-pub mod jcli_transaction;
 pub mod jcli;
 pub mod jormungandr;
-pub mod jormungandr_config;
-pub mod networking; // The purpose of this file is to allow cargo correctly detect tests located in subfolders. It acts like lib.rs file.
+pub mod networking;
+
+// The purpose of this file is to allow cargo correctly detect tests located in subfolders. It acts like lib.rs file.
